notification

+++ client/tinyguard/lib/data/datasource/remote/entity/bounding_box_entity.dart
@@ -22,7 +22,7 @@ class BoundingBoxEntity {
       w: json["w"],
       h: json["h"],
       label: json["label"],
-      confidence: json["confidence"],
+      confidence: json["confidence"].toDouble(),
     );
   }
